+++ src/components/Spawn/ExpirationField.tsx
@@ -6,7 +6,6 @@ import DatePicker from "components/DatePicker";
 import { InputLabel } from "components/styles";
 import TimePicker from "components/TimePicker";
 import { size } from "constants/tokens";
-
 import { useDisableSpawnExpirationCheckbox, useSpruceConfig } from "hooks";
 import { MyHost, MyVolume } from "types/spawn";
 import { SectionContainer, SectionLabel } from "./Layout";
